Create fetch.php
Fetches the USD price of XRP, fetches the exchange rate to CAD, performs multiplication, and then displays the price of XRP in CAD.  

To apply this for your own legal jurisdiction:

Find your Central Bank's RSS feed which states the daily Close rates relative to the USD.  Get the page's contents and use preg_match to find the string you want to search.  Convert the array into a string using json_encode.  Search the string for the price data using substr an store it in a variable to display it to the user.